Add completed task functionality
Tasks can be marked as completed, where text on the task card will be crossed out, and the task will not reload on page load. However, the load completed tasks button will bring those completed tasks to the top of the list. Left to do is making this completed button look good in a media query for phones.